requirements.txt
python-binance

